Changed allometric growth parameters from float to decimal

+++ db/migrate/20150219142625_create_parameters.rb
@@ -3,8 +3,8 @@ class CreateParameters < ActiveRecord::Migration
     create_table :parameters do |t|
       t.float :length_at_capture
       t.float :length_at_maturity
-      t.float :wlen_a
-      t.float :wlen_b
+      t.decimal :wlen_a
+      t.decimal :wlen_b
       t.integer :animal_id
       t.timestamps
     end
